fix: simplify submodule helpers
GitSubmoduleStatus status handling was very complex,
status was not always evaluated (as it requires CommitData).

* Split SubmoduleResources from LocalizationHelpers
* Clean up SubmoduleHelpers name
* Avoid GetMergeBase() call when setting submodule status.
* Set submodule status always after parsing. It was not set in all
situations. This could cause commitdata to be evaluated twice if
new/old commit was done in separate branches.
* Prepare methods in IGitModule and GitSubmoduleStatus to be removed.
