+++ src/app/GitUI/CommandsDialogs/FormCommit.cs
@@ -1874,8 +1874,6 @@ public sealed partial class FormCommit : GitModuleForm
                         {
                             continue;
                         }
-
-                        gitSubmoduleStatus.Status = SubmoduleStatus.Unknown;
                     }
 
                     (GitRevision _, GitRevision indexRev, GitRevision workTreeRev) = GetHeadRevisions();
