+++ src/app/GitUI/Editor/FileViewer.cs
@@ -714,7 +714,7 @@ public partial class FileViewer : GitModuleControl
                 file.IsSubmodule,
                 getImage: () => ThreadHelper.JoinableTaskFactory.Run(GetImageAsync),
                 getFileText: GetFileText,
-                getSubmoduleText: () => LocalizationHelpers.GetSubmoduleText(Module, file.Name.TrimEnd('/'), blobId.ToString()),
+                getSubmoduleText: () => SubmoduleResources.GetSubmoduleText(Module, file.Name.TrimEnd('/'), blobId.ToString()),
                 item: item,
                 line: line,
                 openWithDifftool: openWithDifftool,
@@ -799,7 +799,7 @@ public partial class FileViewer : GitModuleControl
                 isSubmodule,
                 getImage: GetImage,
                 getFileText: GetFileText,
-                getSubmoduleText: () => LocalizationHelpers.GetSubmoduleText(Module, fileName.TrimEnd('/'), ""),
+                getSubmoduleText: () => SubmoduleResources.GetSubmoduleText(Module, fileName.TrimEnd('/'), ""),
                 item: item,
                 line: line,
                 openWithDifftool,
