+++ src/app/GitUI/UserControls/FileStatusList.cs
@@ -1568,8 +1568,8 @@ public sealed partial class FileStatusList : GitModuleControl
             (SubmoduleStatus.NewerTime, false) => nameof(Images.SubmoduleRevisionSemiUp),
             (SubmoduleStatus.OlderTime, true) => nameof(Images.SubmoduleRevisionSemiDownDirty),
             (SubmoduleStatus.OlderTime, false) => nameof(Images.SubmoduleRevisionSemiDown),
-            (_, true) => nameof(Images.SubmoduleDirty),
-            (_, false) => nameof(Images.FolderSubmodule),
+            (SubmoduleStatus.SameTime, false) => nameof(Images.FolderSubmodule),
+            _ => nameof(Images.SubmoduleDirty),
         };
     }
